style: separate folder for SidebarItem
