+++ wondrous-app/components/Common/Sidebar/Entity/List.tsx
@@ -8,8 +8,8 @@ import PieChartIcon from 'components/Common/Sidebar/Common/icons/pieChart.svg';
 import ShowChartIcon from 'components/Common/Sidebar/Common/icons/showChart.svg';
 import StackIcon from 'components/Common/Sidebar/Common/icons/stack.svg';
 import StartIcon from 'components/Common/Sidebar/Common/icons/star.svg';
-import Item from 'components/Common/Sidebar/Common/Item';
 import { Label, ListWrapper } from 'components/Common/Sidebar/Common/styles';
+import Item from 'components/Common/SidebarItem';
 import { GET_ORG_SIDEBAR_COUNT, GET_POD_SIDEBAR_COUNT } from 'graphql/queries';
 import { useRouter } from 'next/router';
 import { ENTITIES_TYPES } from 'utils/constants';
