+++ wondrous-app/components/Common/Sidebar/Resources.tsx
@@ -9,6 +9,7 @@ import {
 } from 'components/Common/Sidebar/Common/styles';
 import BackButton from 'components/Common/SidebarBackButton';
 import CollapseExpandButton from 'components/Common/SidebarCollapseButton';
+import Item from 'components/Common/SidebarItem';
 import useSideBar from 'hooks/useSideBar';
 import { shuffle } from 'lodash';
 import { ColorTypes } from 'utils/constants';
@@ -16,7 +17,6 @@ import { useBoards } from 'utils/hooks';
 
 import { useCanEdit } from './Common/hooks';
 import FolderIcon from './Common/icons/folder.svg';
-import Item from './Common/Item';
 
 const randomColors = shuffle(Object.values(ColorTypes));
